Add Todo page and update site name
